requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
decorator==4.4.0
docker-pycreds==0.4.0
elementpath==1.1.8
future==0.18.2
gitdb==4.0.5
GitPython==3.1.7
gql==0.2.0
graphql-core==1.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
kiwisolver==1.2.0
matplotlib==3.3.1
moviepy==1.0.3
networkx==2.2
numpy==1.18.1
nvidia-ml-py3==7.352.0
opencv-python==4.4.0.42
pandas==1.1.1
pathtools==0.1.2
Pillow==7.2.0
proglog==0.1.9
promise==2.3
psutil==5.6.3
py-trees==0.8.3
pydot==1.4.1
pygame==1.9.6
pyparsing==2.4.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
scikit-image==0.17.2
scipy==1.5.2
sentry-sdk==0.16.5
Shapely==1.6.4.post2
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
tifffile==2020.8.13
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
wandb==0.9.5
watchdog==0.10.3
xmlschema==1.0.13
